RAFT raft.py moved checkpointing suff and removed hf fingerprint code
